feat(fabs): move flower-menu into fabs category
