+++ content/docs/fabs/flower-menu.mdx
@@ -4,7 +4,7 @@ description: A circular flower menu with several icons and a central close butto
 author: arjuncodess
 ---
 
-<ComponentPreview name="list-flower-menu--primary" />
+<ComponentPreview name="fabs-flower-menu--primary" />
 
 ## Installation
 
@@ -12,17 +12,17 @@ author: arjuncodess
 
 <Step>Run the following command</Step>
 
-It will create a new file `flower-menu.tsx` inside the `components/animata/list` directory.
+It will create a new file `flower-menu.tsx` inside the `components/animata/fabs` directory.
 
 ```bash
-mkdir -p components/animata/list && touch components/animata/list/flower-menu.tsx
+mkdir -p components/animata/fabs && touch components/animata/fabs/flower-menu.tsx
 ```
 
 <Step>Paste the code</Step>
 
 Open the newly created file and paste the following code:
 
-```jsx file=<rootDir>/animata/list/flower-menu.tsx
+```jsx file=<rootDir>/animata/fabs/flower-menu.tsx
 
 ```
